requirements.txt
aiosqlite==0.20.0
algokit-utils==2.3.0
altgraph==0.17.4
anyio==4.4.0
argcomplete==3.4.0
azure-core==1.30.2
azure-servicebus==7.12.2
bidict==0.23.1
blinker==1.8.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
customtkinter==5.2.2
darkdetect==0.8.0
Deprecated==1.2.14
dnspython==2.6.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SocketIO==5.3.6
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.7
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
msgpack==1.0.8
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pefile==2024.8.26
pillow==10.4.0
pipx==1.6.0
platformdirs==4.2.2
psutil==6.0.0
py-algorand-sdk==2.6.1
pycparser==2.22
pycryptodomex==3.20.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pymongo==4.8.0
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-engineio==4.9.1
python-socketio==5.11.3
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.3
requests==2.32.3
rfc3986==1.5.0
sanyam_algosdk_2315==0.0.6
scapy==2.5.0
scikit-learn==1.4.1.post1
scipy==1.14.0
screeninfo==0.8.1
setuptools==74.0.0
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
threadpoolctl==3.5.0
tk==0.1.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
userpath==1.9.2
websocket-client==1.8.0
Werkzeug==3.0.3
WMI==1.5.1
wrapt==1.16.0
wsproto==1.2.0
